requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py38h924ce5b_2
cycler=0.10.0=py_2
dbus=1.13.16=hb2f20db_0
expat=2.2.9=he1b5a44_2
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.3=he06d7ca_0
glib=2.66.1=h92f7085_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=hf484d3e_1000
intel-openmp=2020.2=254
joblib=0.17.0=py_0
jpeg=9d=h516909a_0
kiwisolver=1.2.0=py38hbf85e49_1
lcms2=2.11=hbd6801e_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hed695b0_2
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=hc7e4089_6
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he1b5a44_3
matplotlib=3.3.2=py38h32f6830_1
matplotlib-base=3.3.2=py38h4d1ce4f_1
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1h=h516909a_0
pcre=8.44=he1b5a44_0
pillow=8.0.0=py38h9776b28_0
pip=20.2.3=py38_0
pthread-stubs=0.4=h14c3975_1001
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python-mnist=0.7=pypi_0
python_abi=3.8=1_cp38
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.5.2=py38h0b6359f_0
setuptools=50.3.0=py38hb0f4dca_1
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tabulate=0.8.7=pyh9f0ad1d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tornado=6.0.4=py38h1e0a361_2
wheel=0.35.1=py_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h6597ccf_2
